refactor(topics): Refactor to use two models, Topic and TopicPool to represent many to many relationship
