--- admin_tools/models/__init__.py
@@ -1,1 +1,0 @@
-from .topic_pool_link import TopicPoolLink
