+++ topics/models/__init__.py
@@ -1,0 +1,2 @@
+from .topic import Topic
+from .pool import TopicPool
